pyproject.toml
[tool.poetry]
name = "Sayaka_Bot"
version = "0.1.0"
description = "Sayaka_Bot"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0-beta.1"
nonebot-adapter-onebot = "^2.2.1"
# nonebot-plugin-setu-collection = "^0.1.4"
retry = "^0.9.2"
autopep8 = "^2.0.1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = []
plugin_dirs = ["src/plugins"]
adapters = [{name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11", project_link = "nonebot-adapter-onebot", desc = "OneBot V11 协议"}]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
